+++ src/tools.hpp
@@ -9,7 +9,9 @@ namespace tools {
     void connectObjects(EditorUI* editorUI, CCObject* sender);
     void fillObjects(EditorUI* editorUI, CCObject* sender);
     void replaceObjectType(EditorUI* editorUI, CCObject* sender);
+    void replicateObject(EditorUI* editorUI, CCObject* sender);
     void circleObjects(EditorUI* editorUI, CCObject* sender);
     void connectTrigger(EditorUI* editorUI, CCObject* sender);
     void disconnectTrigger(EditorUI* editorUI, CCObject* sender);
+    void spawnedTrigger(EditorUI* editorUI, CCObject* sender);
 }